chatbot/requirements.txt
langchain-community
langchain-core
langchain-text-splitters
langchain[anthropic,openai]
langgraph
markdown
python-dotenv
unstructured
